--- metadata/creation_process/100.md
@@ -1,37 +1,0 @@
-# prompt-1
-
-# 指示
-以下の条件を全て満たすような、大学機関におけるIR活動で使用する合成データ（synthetic data）を生成するPythonプログラムを生成してください。
-
-# 条件
-1. ヘッダーは、連番のID（S{A:B:C:D:E}{0000-4999}）、氏名（名字+名前）、性別（男または女）、所属学部（文学部、理工学部、情報学部、経済学部、理学部）、各科目（古典、物理、化学、経済、英語、数学、情報）の100点満点の点数とする。
-2. 各学部は1000名ずつで、文学部ならIDは0000-0999、理工学部ならIDは1000-1999、情報学部ならIDは2000-2999、経済学部ならIDは3000-3999、理学部ならIDは4000-4999とする。
-3. 所属学部に応じて各科目の素点を上下させる。
-各素点は整数にするが、学部に応じてある程度の傾向を設定する。
-4. 氏名は日本人の名前にする。
-その際には、fakerライブラリを使うこと。氏名と性別から矛盾ないように設定する。
-5. その他、統計的な処理をしたときに差があるようにデータを調整して生成する。
-6. 出力先は"../../datasets/100.csv"として書き出す。
-7. PEP8に則ったPythonプログラムである。
-
-# prompt-2
-
-あなたはPythonのデータ分析に精通しているエンジニアである。
-"../../datasets/100.csv"というファイルが以下のような形式で渡された。
-
-ID,氏名,性別,所属学部,古典,物理,化学,経済,英語,数学,情報
-S0000,鈴木 充,男,文学部,95,57,60,45,85,47,63
-S0001,佐藤 和也,男,文学部,92,49,58,60,79,65,53
-S0002,吉田 太郎,男,文学部,99,52,58,48,87,60,64
-S0003,阿部 香織,女,文学部,94,64,47,49,88,51,55
-S0004,吉田 あすか,女,文学部,93,45,63,53,78,56,51
-...
-
-このデータをよしなに分析して、以下の項目について結果を教えてください。
-
-1. 各学部の学生数と成績平均
-2. 性別ごとの成績平均
-3. 英語、数学、情報リテラシーの成績分布をヒストグラムで可視化
-4. 各学部や各科目の違いによる統計的な差
-
-分析の結果を文章として報告してください。また、可視化したヒストグラムを表示してください。よろしくお願いします。
